requirements.txt
pandas
numpy
scikit-learn
matplotlib
scikit-surprise
lightfm
streamlit